+++ database/migrations/2024_04_01_162524_create_fornos_table.php
@@ -18,11 +18,11 @@ class CreateFornosTable extends Migration
         $table->integer('user_id')->index()->nullable();
         $table->integer('produto_id')->index()->nullable();
         $table->integer('extrusora_id')->index()->nullable();
-        $table->integer('peso')->nullable();
-        $table->integer('dim_externa')->nullable();
-        $table->integer('dim_parede')->nullable();
-        $table->integer('umidade')->nullable();
-        $table->integer('resistencia')->nullable();
+        $table->double('peso',8,2)->nullable();
+        $table->double('dim_externa',8,2)->nullable();
+        $table->double('dim_parede',8,2)->nullable();
+        $table->double('umidade',8,2)->nullable();
+        $table->double('resistencia',8,2)->nullable();
         $table->string('residuo',100)->nullable();
 
         $table->softDeletes();
