+++ database/migrations/2024_05_06_181039_add_medidaforno.php
@@ -10,8 +10,8 @@ class AddMedidaforno extends Migration
     public function up()
     {
         Schema::table('forno', function (Blueprint $table) {
-            $table->integer('absorcao')->nullable();
-            $table->integer('cod_residuo')->nullable();
+            $table->double('absorcao',8,2)->nullable();
+            $table->double('cod_residuo',8,2)->nullable();
             $table->integer('historico_id')->index()->nullable();
         });
     }
